+++ GFramework.Core.Tests/State/StateMachineTests.cs
@@ -1,4 +1,3 @@
-using System.Reflection;
 using GFramework.Core.Abstractions.State;
 using GFramework.Core.State;
 using NUnit.Framework;
@@ -366,239 +365,5 @@ public class StateMachineTests
         Assert.That(state2.ExitCalled, Is.True);
         Assert.That(state3.EnterCalled, Is.True);
         Assert.That(state3.ExitCalled, Is.False);
-    }
-}
-
-/// <summary>
-///     测试状态类V2版本，实现IState接口用于测试
-/// </summary>
-public sealed class TestStateV2 : IState
-{
-    public bool AllowTransition { get; set; } = true;
-    public bool EnterCalled { get; private set; }
-    public bool ExitCalled { get; private set; }
-    public int EnterCallCount { get; private set; }
-    public int ExitCallCount { get; private set; }
-    public IState? EnterFrom { get; private set; }
-    public IState? ExitTo { get; private set; }
-
-    /// <summary>
-    ///     进入状态时的回调方法
-    /// </summary>
-    /// <param name="from">从哪个状态进入</param>
-    public void OnEnter(IState? from)
-    {
-        EnterCalled = true;
-        EnterCallCount++;
-        EnterFrom = from;
-    }
-
-    /// <summary>
-    ///     离开状态时的回调方法
-    /// </summary>
-    /// <param name="to">离开到哪个状态</param>
-    public void OnExit(IState? to)
-    {
-        ExitCalled = true;
-        ExitCallCount++;
-        ExitTo = to;
-    }
-
-    /// <summary>
-    ///     判断是否可以转换到目标状态
-    /// </summary>
-    /// <param name="target">目标状态</param>
-    /// <returns>是否允许转换</returns>
-    public bool CanTransitionTo(IState target)
-    {
-        return AllowTransition;
-    }
-}
-
-/// <summary>
-///     测试状态类V3版本，实现IState接口用于测试
-/// </summary>
-public sealed class TestStateV3 : IState
-{
-    public bool EnterCalled { get; private set; }
-    public bool ExitCalled { get; private set; }
-    public int EnterCallCount { get; private set; }
-    public int ExitCallCount { get; private set; }
-    public IState? EnterFrom { get; private set; }
-    public IState? ExitTo { get; private set; }
-
-    /// <summary>
-    ///     进入状态时的回调方法
-    /// </summary>
-    /// <param name="from">从哪个状态进入</param>
-    public void OnEnter(IState? from)
-    {
-        EnterCalled = true;
-        EnterCallCount++;
-        EnterFrom = from;
-    }
-
-    /// <summary>
-    ///     离开状态时的回调方法
-    /// </summary>
-    /// <param name="to">离开到哪个状态</param>
-    public void OnExit(IState? to)
-    {
-        ExitCalled = true;
-        ExitCallCount++;
-        ExitTo = to;
-    }
-
-    /// <summary>
-    ///     判断是否可以转换到目标状态
-    /// </summary>
-    /// <param name="target">目标状态</param>
-    /// <returns>是否允许转换</returns>
-    public bool CanTransitionTo(IState target)
-    {
-        return true;
-    }
-}
-
-/// <summary>
-///     测试状态类V4版本，实现IState接口用于测试
-/// </summary>
-public sealed class TestStateV4 : IState
-{
-    public bool EnterCalled { get; private set; }
-    public bool ExitCalled { get; private set; }
-    public int EnterCallCount { get; private set; }
-    public int ExitCallCount { get; private set; }
-    public IState? EnterFrom { get; private set; }
-    public IState? ExitTo { get; private set; }
-
-    /// <summary>
-    ///     进入状态时的回调方法
-    /// </summary>
-    /// <param name="from">从哪个状态进入</param>
-    public void OnEnter(IState? from)
-    {
-        EnterCalled = true;
-        EnterCallCount++;
-        EnterFrom = from;
-    }
-
-    /// <summary>
-    ///     离开状态时的回调方法
-    /// </summary>
-    /// <param name="to">离开到哪个状态</param>
-    public void OnExit(IState? to)
-    {
-        ExitCalled = true;
-        ExitCallCount++;
-        ExitTo = to;
-    }
-
-    /// <summary>
-    ///     判断是否可以转换到目标状态
-    /// </summary>
-    /// <param name="target">目标状态</param>
-    /// <returns>是否允许转换</returns>
-    public bool CanTransitionTo(IState target)
-    {
-        return true;
-    }
-}
-
-/// <summary>
-///     异步测试状态类，同时实现IState和IAsyncState接口用于测试异步状态功能
-/// </summary>
-public sealed class TestAsyncState : IState, IAsyncState
-{
-    public bool AllowTransition { get; set; } = true;
-    public bool EnterCalled { get; private set; }
-    public bool ExitCalled { get; private set; }
-    public int EnterCallCount { get; private set; }
-    public int ExitCallCount { get; private set; }
-    public IState? EnterFrom { get; private set; }
-    public IState? ExitTo { get; private set; }
-    public int CanTransitionToCallCount { get; private set; }
-
-    /// <summary>
-    ///     异步进入状态时的回调方法
-    /// </summary>
-    /// <param name="from">从哪个状态进入</param>
-    public async Task OnEnterAsync(IState? from)
-    {
-        await Task.Delay(1).ConfigureAwait(false);
-        EnterCalled = true;
-        EnterCallCount++;
-        EnterFrom = from;
-    }
-
-    /// <summary>
-    ///     异步离开状态时的回调方法
-    /// </summary>
-    /// <param name="to">离开到哪个状态</param>
-    public async Task OnExitAsync(IState? to)
-    {
-        await Task.Delay(1).ConfigureAwait(false);
-        ExitCalled = true;
-        ExitCallCount++;
-        ExitTo = to;
-    }
-
-    /// <summary>
-    ///     异步判断是否可以转换到目标状态
-    /// </summary>
-    /// <param name="target">目标状态</param>
-    /// <returns>是否允许转换</returns>
-    public async Task<bool> CanTransitionToAsync(IState target)
-    {
-        await Task.Delay(1).ConfigureAwait(false);
-        CanTransitionToCallCount++;
-        return AllowTransition;
-    }
-
-    /// <summary>
-    ///     进入状态时的回调方法（同步版本，抛出异常表示不应被调用）
-    /// </summary>
-    /// <param name="from">从哪个状态进入</param>
-    public void OnEnter(IState? from)
-    {
-        throw new InvalidOperationException("Sync OnEnter should not be called for async state");
-    }
-
-    /// <summary>
-    ///     离开状态时的回调方法（同步版本，抛出异常表示不应被调用）
-    /// </summary>
-    /// <param name="to">离开到哪个状态</param>
-    public void OnExit(IState? to)
-    {
-        throw new InvalidOperationException("Sync OnExit should not be called for async state");
-    }
-
-    /// <summary>
-    ///     判断是否可以转换到目标状态（同步版本，抛出异常表示不应被调用）
-    /// </summary>
-    /// <param name="target">目标状态</param>
-    /// <returns>是否允许转换</returns>
-    public bool CanTransitionTo(IState target)
-    {
-        throw new InvalidOperationException("Sync CanTransitionTo should not be called for async state");
-    }
-}
-
-/// <summary>
-///     状态机扩展方法类
-/// </summary>
-public static class StateMachineExtensions
-{
-    /// <summary>
-    ///     检查状态机是否包含指定类型的状态
-    /// </summary>
-    /// <typeparam name="T">要检查的状态类型</typeparam>
-    /// <param name="stateMachine">状态机实例</param>
-    /// <returns>如果包含指定类型的状态则返回true，否则返回false</returns>
-    public static bool ContainsState<T>(this StateMachine stateMachine) where T : IState
-    {
-        return stateMachine.GetType().GetField("States", BindingFlags.NonPublic | BindingFlags.Instance)?
-                   .GetValue(stateMachine) is Dictionary<Type, IState> states &&
-               states.ContainsKey(typeof(T));
     }
 }
